Refactor: Remove debug logs and consolidate auth middleware
